sandbox/requirements.txt
asn1crypto==1.3.0
bit==0.6.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
coincurve==13.0.0
cytoolz==0.10.1
ecdsa==0.15
eciespy==0.3.4
eth-hash==0.2.0
eth-keys==0.3.1
eth-typing==2.2.1
eth-utils==1.8.4
Flask==1.1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pycparser==2.19
pycryptodome==3.9.7
requests==2.23.0
six==1.14.0
toolz==0.10.0
transmissionrpc==0.11
urllib3==1.25.8
Werkzeug==1.0.0
